Feat: adjust controller update and delete with redis

+++ app/Http/Controllers/EselonController.php
@@ -3,7 +3,9 @@
 namespace App\Http\Controllers;
 
 use App\Models\Eselon;
+use App\Models\PendingAction;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class EselonController extends Controller
 {
@@ -38,11 +40,27 @@ class EselonController extends Controller
      */
     public function update(Request $request)
     {
+        $eselon = Eselon::findOrFail($request->id);
+
         $validatedData = $request->validate([
             'name' => 'required|string|max:255',
         ]);
 
-        $eselon = Eselon::whereId($request->id)->firstOrFail();
+        if (Auth::user()->role_id == 2) {
+            $updateData = [
+                'id' => $eselon->id,
+                'validated_data' => $validatedData,
+                'type' => 'Eselon',
+                'requested_by' => Auth::user()->id,
+                'requested_at' => now(),
+            ];
+
+            PendingAction::savePendingAction('update', $eselon->id, $updateData);
+
+            return redirect()->back()
+                ->with('success', 'Update request has been sent for approval.');
+        }
+
         $eselon->update($validatedData);
 
         return redirect()->back()->with('success', 'Eselon updated successfully');
@@ -53,7 +71,21 @@ class EselonController extends Controller
      */
     public function destroy(Request $request)
     {
-        $eselon = Eselon::whereId($request->id)->firstOrFail();
+        $eselon = Eselon::findOrFail($request->id);
+        if (Auth::user()->role_id == 2) {
+            $deleteData = [
+                'id' => $eselon->id,
+                'name' => $eselon->name,
+                'type' => 'Eselon',
+                'requested_by' => Auth::user()->id,
+                'requested_at' => now(),
+            ];
+
+            PendingAction::savePendingAction('delete', $eselon->id, $deleteData);
+
+            return redirect()->back()
+                ->with('success', 'Delete request has been sent for approval.');
+        }
         $eselon->delete();
 
         return redirect()->back()->with('success', 'Eselon deleted successfully');
